backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
pydantic
pyyaml
pytest
llama-cpp-python>=0.3.16,<0.4
pandas
openpyxl
python-multipart
pillow
